backend/requirements.txt
# Django and related packages
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.2.0
django-filter==23.3
djangorestframework-simplejwt==5.3.0

# Database
psycopg2-binary==2.9.9
dj-database-url==2.1.0

# API Integration
fastapi==0.104.1
uvicorn==0.24.0
httpx==0.25.1
openai==1.2.4
google-generativeai==0.2.1

# Security
python-dotenv==1.0.0
cryptography==41.0.4

# Utilities
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
python-dateutil==2.8.2
setuptools==80.3.1
django-cors-headers